refactor(frontend)(#1422): split App.tsx + BlockNode.tsx into <=500 LOC modules
App.tsx (1379 LOC -> 494) and BlockNode.tsx (1176 LOC -> 295) are split
into focused modules so every produced file fits the eslint max-lines
cap (500 LOC, the same cap the #1422 god-file refactor umbrella will
ultimately enforce once Wave 4 removes the size waivers).

App.tsx delegates to hooks + presentation components under
frontend/src/App.parts/:
- AppLevelMergeFlow.tsx        (ADR-039 §3.5 mount)
- InteractiveModals.tsx        (DataRouter / PairEditor prompts)
- ProjectWorkspace.tsx         (three-column ResizablePanelGroup tree)
- useAppKeyboardShortcuts.ts   (global Ctrl-S / Ctrl-Z listener)
- useAppLifecycleEffects.ts    (boot + workflow autosave + tab sync)
- useBottomPanelControls.ts    (panel imperative handle + ref)
- useCanvasHandlers.ts         (palette add, edge connect, view source)
- useFileTabsAutosave.ts       (ADR-036 §3.9 per-tab debouncer)
- useProjectActions.ts         (project / workflow / file CRUD)
- useWorkflowExecutionActions  (run / pause / resume / cancel / per-block)
- useWorkflowSync.ts           (refreshProjects / refreshBlocks / save)

BlockNode.tsx delegates to focused modules under
frontend/src/components/nodes/BlockNode.parts/:
- badgeStyles.ts               (category icons + StatusBadge styles)
- StatusBadge.tsx              (status pill)
- ErrorMessage.tsx             (inline truncated error)
- PausedToast.tsx              (AppBlock paused chip)
- FileBrowserModal.tsx         (fallback path picker)
- inlineConfigHelpers.ts       (ConfigProperty + capability helpers)
- InlineCapabilitySelector.tsx (Format dropdown)
- InlineConfigField.tsx        (enum/bool/number/textarea discriminator)
- InlineTextInputField.tsx     (default text input with #710 caret save)
- PortHandles.tsx              (Handle rendering + ADR-029 variadic +/-)

BlockNode.test.tsx (985 LOC) is split into themed test files under
frontend/src/components/nodes/__tests__/BlockNode/:
- test-utils.tsx               (shared fixtures + api mock)
- capabilities.test.tsx        (ADR-043 + #1307 + hidden direction)
- ports.test.tsx               (ADR-028 §D4 dynamic ports + #467)
- errors.test.tsx              (#422 inline error display)
- caret.test.tsx               (#710 caret preservation + audit follow-up)
- nativeDialog.test.tsx        (#678 status-aware fallback)
- hooks1420.test.tsx           (#1420 hook order regression coverage)
- inlineConfigHelpers.test.ts  (14 new pure-unit cases)

Behavior is preserved verbatim. Wave 1 (#1420/#1421) discipline survives
the split intact: InlineTextInputField stays a top-level component so
its hook chain is independent of the discriminator's early returns;
every useCallback in App.tsx retains its full reactive dep list, and
every useEffect's dep array is exhaustive (or carries the same inline
disable + rationale as the pre-split version). react-hooks/rules-of-hooks
and react-hooks/exhaustive-deps both pass with zero warnings on
App.tsx + BlockNode.tsx + the new files.

The GOD_FILE_SIZE_WAIVERS block in frontend/eslint.config.js is
untouched in this PR; the manager removes App.tsx/BlockNode.tsx/
BlockNode.test.tsx entries in Wave 4 after all three W2 PRs land.

Tests: 494 passed | 13 skipped (507 total) across 51 test files (was
481 before the split — 14 new in inlineConfigHelpers.test.ts). Lint /
typecheck / format / build / full_audit / sentrux all clean.

Gate-Record: .workflow/records/1422-god-app-blocknode.json
Task-Kind: refactor
Issue: #1422
