+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Changed
 
+- [#1422] Split `frontend/src/App.tsx` (1379 LOC) and `frontend/src/components/nodes/BlockNode.tsx` (1176 LOC) into focused modules so every produced file is ≤ 500 LOC, per the #1422 god-file refactor umbrella. **App.tsx** (now 494 LOC) delegates large concerns to hooks under `frontend/src/App.parts/`: `useAppKeyboardShortcuts` (global Ctrl-S/Ctrl-Z listener), `useAppLifecycleEffects` (boot + workflow autosave + tab snapshot sync), `useBottomPanelControls` (`ResizablePanel` imperative handle + `handleNodeSelect`/`handleErrorClick`/`handleBottomTabChange`/`handleCanvasPaneClick`), `useCanvasHandlers` (palette add, edge connect, View source, Save router), `useFileTabsAutosave` (ADR-036 §3.9 per-tab debouncer), `useProjectActions` (project/workflow/file CRUD), `useWorkflowExecutionActions` (run/pause/resume/cancel/start/per-block), `useWorkflowSync` (refresh + saveWorkflow + saveWorkflowAs); plus three presentation components — `AppLevelMergeFlow.tsx`, `InteractiveModals.tsx`, `ProjectWorkspace.tsx` (the three-column `ResizablePanelGroup` tree). **BlockNode.tsx** (now 295 LOC) delegates to focused modules under `frontend/src/components/nodes/BlockNode.parts/`: `badgeStyles.ts` (category-icon + status-badge style table), `StatusBadge.tsx`, `ErrorMessage.tsx`, `PausedToast.tsx`, `FileBrowserModal.tsx`, `inlineConfigHelpers.ts` (`ConfigProperty` + capability helpers), `InlineCapabilitySelector.tsx`, `InlineConfigField.tsx` (enum/boolean/number/textarea discriminator), `InlineTextInputField.tsx` (default text-input branch with #710 caret-preservation), `PortHandles.tsx` (input/output `Handle`s + ADR-029 variadic add/remove). The original `BlockNode.test.tsx` (985 LOC) is split into themed files under `frontend/src/components/nodes/__tests__/BlockNode/`: `test-utils.tsx` (shared fixtures + api mock + `reactNativeInputChange`), `capabilities.test.tsx` (ADR-043 + #1307 + hidden-direction), `ports.test.tsx` (ADR-028 §D4 dynamic ports + #467 Browse removal), `errors.test.tsx` (#422 inline error), `caret.test.tsx` (#710 caret preservation + audit follow-up), `nativeDialog.test.tsx` (#678 fallback), `hooks1420.test.tsx` (#1420 `InlineTextInputField` hook order regression coverage), and `inlineConfigHelpers.test.ts` (14 new pure-unit cases over the extracted helpers); `BlockNode.test.tsx` keeps only the sanity smoke tests. Behavior is preserved verbatim — Wave 1's #1420 fix (`InlineTextInputField` extraction at the top of every component's hook chain) and #1421 fixes (`useCallback` wrappers + exhaustive-deps repair on `App.tsx`) survive the split intact; `react-hooks/rules-of-hooks` and `react-hooks/exhaustive-deps` both still pass with zero warnings on `App.tsx`/`BlockNode.tsx` and all extracted files. The `GOD_FILE_SIZE_WAIVERS` block in `frontend/eslint.config.js` is untouched in this PR (manager removes `App.tsx`/`BlockNode.tsx`/`BlockNode.test.tsx` entries in Wave 4 after all three W2 PRs land). 494 frontend tests + 13 skipped pass; lint/typecheck/format/build clean. (@claude, 2026-05-22, branch: refactor/issue-1422/god-app-blocknode)
+
 - [#1408] Downgrade Sentrux from a hard gate to an advisory signal (ADR-042 Addendum 3). The Sentrux binary is not installable in many contributor and AI-agent execution environments (distributed as Rust source under the Claude marketplace; no prebuilt binary), so every source-/workflow-/architecture-/governance-touching PR was hitting `gate-record.sentrux.missing` at the local pre-push hook and at CI's workflow-gate, requiring an `admin-approved:ai-override` label purely for tooling reasons (worked example: PR #1406 shipped with the override label even though no `src/scistudio/` was touched). Two changes: (1) `src/scistudio/qa/governance/gate_record.py::validate_gate_record` no longer emits `gate-record.sentrux.missing` — missing, skipped, and unknown sentrux evidence are all allowed. The narrower `gate-record.sentrux.not-passing` finding is kept and tightened to fire only on recorded `status="fail"` (status="unknown" and status="skipped" no longer count) so the gate's remaining scope is the case it was actually designed for: a developer who ran sentrux, observed a real failure, and is pushing anyway. The override label and `--bypass-label` flag still work for justified exceptions. (2) `.github/workflows/workflow-gate.yml` calls `sentrux_gate.verify_free_tier_claims` unchanged but converts findings to `::warning::` annotations instead of adding them to the blocking reports list — the workflow-gate job never fails on sentrux alone, and reviewers read the advisories in PR check details. The supersession is narrow: only Addendum 1 §D1 (sentrux required as gate evidence) and §D5 (free-tier-honesty rule's blocking semantics) are affected. §D2 committed gate records, §D3 six-stage gate, §D7 implementation tests required, §D8 legacy gate removal, §D9 generated artifacts, §D10 AI gate CLI requirement, and §D11/§D12 architecture frontmatter / drift checks are unaffected. The Sentrux tool, `SentruxEvidence` data model, `gate_record sentrux` CLI subcommand, `verify_free_tier_claims`, and `.sentrux/rules.toml` all continue to exist; only the enforcement is downgraded. Tests reflect the inverted semantics: `tests/qa/test_gate_record_ci.py::test_sentrux_required_for_applicable_changes` renamed to `test_sentrux_missing_evidence_is_allowed_for_applicable_changes` (asserts missing does NOT block); two new tests `test_sentrux_skipped_evidence_is_allowed_for_applicable_changes` and `test_sentrux_recorded_failure_still_blocks_locally` cover the new threshold. Existing `test_sentrux_evidence_rejects_*` tests in `tests/qa/test_gate_record.py` retained unchanged because the `SentruxEvidence` validation model itself is unchanged (Pro-required claims and impossible rule counts still rejected at parse time). (@agent, 2026-05-21, branch: feat/sentrux-advisory)
 ### Fixed
